open the task editor in a tab as an alternative to the modal
