+++ README.md
@@ -136,6 +136,7 @@ Task hierarchy and dependencies don't resolve on the TaskNotes side (it uses pro
 |---|---|
 | Projects folder | Vault folder where project and task files are stored |
 | Default view | Table, Gantt, or Kanban |
+| Open tasks in | Modal or tab. On tab, task notes open in the task editor instead of Obsidian's. |
 | Gantt granularity | Default timeline scale (day / week / month / quarter) |
 | Gantt week labels | Week number, date range, or both |
 | Due date notifications | Reminders N days before due dates |
