+++ src/modals/TaskModal.ts
@@ -1,602 +1,43 @@
-import {
-  App,
-  ButtonComponent,
-  Component,
-  ExtraButtonComponent,
-  Menu,
-  Modal,
-  MarkdownRenderer,
-  Notice,
-  setIcon,
-  setTooltip
-} from 'obsidian'
+import { App, Modal } from 'obsidian'
 import type PMPlugin from '../main'
-import { type Project, type Task, makeTask } from '../types'
-import { flattenTasks } from '../store/TaskTreeOps'
-import { TaskFileNameConflictError } from '../store'
-import { safeAsync, getDefaultStatusId, getDefaultPriorityId, getPriorityConfig } from '../utils'
-import { confirmDialog } from '../ui/ModalFactory'
-import { renderTaskFormFields } from './TaskFormFields'
-import { renderTimeTrackingPanel } from './TimeTrackingPanel'
-import { renderSubtasksPanel } from './SubtasksPanel'
-import { NoteLinkSuggest } from './NoteLinkSuggest'
+import type { Project, Task } from '../types'
+import { TaskEditor } from './TaskEditor'
 
 export class TaskModal extends Modal {
-  private task: Task
-  /** The task as it was when the editor opened, to diff against on save. */
-  private original: Task
-  private isNew: boolean
-  private originalParentId: string | null
-  private cancelled = false
-  private saved = false
-  private persistPromise: Promise<void> | null = null
-  private noteSuggest: NoteLinkSuggest | null = null
-  private shownExtras = new Set<string>()
-  private saveKeyHandler: ((e: KeyboardEvent) => void) | null = null
+  private editor: TaskEditor
 
   constructor(
     app: App,
-    private plugin: PMPlugin,
-    private project: Project,
+    plugin: PMPlugin,
+    project: Project,
     task: Task | null,
-    private parentId: string | null,
-    private onSave: (task: Task) => void | Promise<void>,
+    parentId: string | null,
+    onSave: (task: Task) => void | Promise<void>,
     defaults?: Partial<Task>
   ) {
     super(app)
-    if (task) {
-      this.task = JSON.parse(JSON.stringify(task)) as Task
-      this.isNew = false
-      if (parentId == null) {
-        const flat = flattenTasks(project.tasks)
-        const entry = flat.find((f) => f.task.id === task.id)
-        this.parentId = entry?.parentId ?? null
-      }
-    } else {
-      const config = plugin.store.configFor(project)
-      this.task = makeTask({
-        status: getDefaultStatusId(config.statuses),
-        priority: getDefaultPriorityId(config.priorities),
-        type: parentId ? 'subtask' : 'task',
-        ...defaults
-      })
-      this.isNew = true
-    }
-    this.original = JSON.parse(JSON.stringify(this.task)) as Task
-    this.originalParentId = this.parentId
+    this.editor = new TaskEditor(
+      app,
+      plugin,
+      project,
+      task,
+      parentId,
+      onSave,
+      { surface: 'modal', close: () => this.close(), keyScopeEl: this.modalEl },
+      defaults
+    )
   }
 
   onOpen(): void {
     const { contentEl } = this
     contentEl.empty()
-    contentEl.addClass('pm-task-modal')
+    contentEl.addClass('pm-task-modal', 'pm-te-surface')
     this.modalEl.addClass('pm-modal', 'pm-modal--task')
-    this.render()
+    this.editor.mount(contentEl)
   }
 
   onClose(): void {
-    if (
-      this.plugin.settings.saveTaskOnClose &&
-      !this.isNew &&
-      !this.cancelled &&
-      !this.saved &&
-      this.task.title.trim()
-    ) {
-      const conflict = this.plugin.store.findTaskFileConflict(this.project, this.task)
-      if (conflict) {
-        new Notice(`Task not saved: a note named "${conflict.fileName}" already exists.`)
-      } else {
-        void this.persistTask()
-      }
-    }
-    if (this.saveKeyHandler) {
-      this.modalEl.removeEventListener('keydown', this.saveKeyHandler)
-      this.saveKeyHandler = null
-    }
-    this.noteSuggest?.destroy()
-    this.noteSuggest = null
+    this.editor.handleClose()
     this.contentEl.empty()
-  }
-
-  private persistTask(): Promise<void> {
-    if (this.persistPromise) return this.persistPromise
-    const p = (async () => {
-      try {
-        await this.runPersist()
-      } catch (err) {
-        this.persistPromise = null
-        throw err
-      }
-    })()
-    this.persistPromise = p
-    return p
-  }
-
-  private async insertAttachments(
-    descArea: HTMLTextAreaElement,
-    items: { blob: Blob; name: string }[],
-    autoResize: () => void
-  ): Promise<void> {
-    for (const { blob, name } of items) {
-      try {
-        const buffer = await blob.arrayBuffer()
-        const file = await this.plugin.store.saveTaskAttachment(this.project, this.task, name, buffer)
-        const snippet = `![[${file.name}]]`
-        descArea.setRangeText(snippet, descArea.selectionStart, descArea.selectionEnd, 'end')
-        this.task.description = descArea.value
-        autoResize()
-      } catch (err) {
-        console.error('Failed to save attachment', err)
-        new Notice('Failed to save attachment')
-      }
-    }
-  }
-
-  /**
-   * Only what changed. The editor works on a copy taken when it opened, so saving the
-   * whole task would write its stale view of every other field back over disk.
-   */
-  private changedFields(): Partial<Task> {
-    const patch: Partial<Task> = {}
-    for (const key of Object.keys(this.task) as (keyof Task)[]) {
-      if (JSON.stringify(this.task[key]) !== JSON.stringify(this.original[key])) {
-        Object.assign(patch, { [key]: this.task[key] })
-      }
-    }
-    return patch
-  }
-
-  private async runPersist(): Promise<void> {
-    if (this.isNew) {
-      await this.plugin.store.insertTask(this.project, this.task, this.parentId)
-    } else {
-      const patch = this.changedFields()
-      const moved = this.parentId !== this.originalParentId
-      if (Object.keys(patch).length === 0 && !moved) return
-      await this.plugin.store.updateTask(this.project, this.task.id, patch)
-      if (moved) await this.plugin.store.moveTask(this.project, this.task.id, this.parentId)
-    }
-    await this.plugin.store.scheduleAfterChange(this.project, this.task.id)
-    await this.onSave(this.task)
-  }
-
-  private openOverflowMenu(anchorEl: HTMLElement): void {
-    const menu = new Menu()
-    if (this.task.filePath) {
-      const filePath = this.task.filePath
-      menu.addItem((item) =>
-        item
-          .setTitle('Open as note')
-          .setIcon('file-text')
-          .onClick(() => {
-            this.saved = false
-            this.cancelled = false
-            this.close()
-            void this.app.workspace.openLinkText(filePath, '', true)
-          })
-      )
-      menu.addSeparator()
-    }
-    if (this.task.archived) {
-      menu.addItem((item) =>
-        item
-          .setTitle('Unarchive')
-          .setIcon('archive-restore')
-          .onClick(
-            safeAsync(async () => {
-              await this.plugin.store.unarchiveTask(this.project, this.task.id)
-              new Notice('Task unarchived')
-              await this.onSave(this.task)
-              this.cancelled = true
-              this.close()
-            })
-          )
-      )
-    } else {
-      menu.addItem((item) =>
-        item
-          .setTitle('Archive')
-          .setIcon('archive')
-          .onClick(
-            safeAsync(async () => {
-              await this.plugin.store.archiveTask(this.project, this.task.id)
-              new Notice('Task archived')
-              await this.onSave(this.task)
-              this.cancelled = true
-              this.close()
-            })
-          )
-      )
-    }
-    menu.addItem((item) =>
-      item
-        .setTitle('Delete')
-        .setIcon('trash-2')
-        .setWarning(true)
-        .onClick(
-          safeAsync(async () => {
-            if (await confirmDialog(this.app, `Delete "${this.task.title}"?`)) {
-              await this.plugin.store.deleteTask(this.project, this.task.id)
-              await this.onSave(this.task)
-              this.cancelled = true
-              this.close()
-            }
-          })
-        )
-    )
-    const rect = anchorEl.getBoundingClientRect()
-    menu.showAtPosition({ x: rect.left, y: rect.bottom + 4 })
-  }
-
-  private render(): void {
-    const { contentEl } = this
-    contentEl.empty()
-
-    const header = contentEl.createDiv('pm-te-header')
-    const prio = getPriorityConfig(this.plugin.store.configFor(this.project).priorities, this.task.priority)
-    if (prio?.color) header.setCssProps({ '--pm-accent-strip': prio.color })
-    const crumb = header.createDiv('pm-te-crumb')
-    if (this.project.icon) {
-      const iconEl = crumb.createSpan({ cls: 'pm-te-crumb-icon' })
-      // project.icon is either an emoji or a Lucide icon name.
-      if (/^[a-z0-9-]+$/.test(this.project.icon)) setIcon(iconEl, this.project.icon)
-      else iconEl.setText(this.project.icon)
-    }
-    crumb.createSpan({ cls: 'pm-te-crumb-name', text: this.project.title })
-    const crumbSep = crumb.createSpan({ cls: 'pm-te-crumb-sep' })
-    setIcon(crumbSep, 'chevron-right')
-    const idEl = crumb.createSpan({ cls: 'pm-te-crumb-id pm-te-copyable', text: this.task.id })
-    setTooltip(idEl, 'Copy task ID')
-    idEl.addEventListener(
-      'click',
-      safeAsync(async () => {
-        await navigator.clipboard.writeText(this.task.id)
-        new Notice('Copied task ID')
-      })
-    )
-
-    header.createDiv('pm-te-header-spacer')
-
-    if (!this.isNew) {
-      const moreBtn = new ExtraButtonComponent(header).setIcon('more-horizontal').setTooltip('More actions')
-      moreBtn.extraSettingsEl.addClass('pm-te-header-btn')
-      moreBtn.onClick(() => this.openOverflowMenu(moreBtn.extraSettingsEl))
-    }
-    const closeBtn = new ExtraButtonComponent(header).setIcon('x').setTooltip('Close')
-    closeBtn.extraSettingsEl.addClass('pm-te-header-btn')
-    closeBtn.onClick(() => {
-      this.cancelled = true
-      this.close()
-    })
-
-    const body = contentEl.createDiv('pm-te-body')
-
-    const titleWrap = body.createDiv('pm-te-title-wrap')
-    const titleInput = titleWrap.createEl('textarea', { cls: 'pm-te-title' })
-    titleInput.rows = 1
-    titleInput.value = this.task.title
-    titleInput.placeholder = 'Task title'
-    titleInput.spellcheck = false
-    const autosizeTitle = () => {
-      titleInput.setCssProps({ '--te-title-height': 'auto' })
-      titleInput.setCssProps({ '--te-title-height': titleInput.scrollHeight + 'px' })
-    }
-    const titleError = titleWrap.createDiv({ cls: 'pm-modal-title-error', attr: { hidden: '' } })
-    const clearTitleError = () => {
-      if (titleError.hasAttribute('hidden')) return
-      titleError.setAttribute('hidden', '')
-      titleError.setText('')
-      titleInput.classList.remove('pm-input-error')
-    }
-    const showTitleError = (message: string) => {
-      titleError.setText(message)
-      titleError.removeAttribute('hidden')
-      titleInput.classList.add('pm-input-error')
-      titleInput.focus()
-      titleInput.select()
-    }
-    titleInput.addEventListener('input', () => {
-      this.task.title = titleInput.value
-      clearTitleError()
-      autosizeTitle()
-    })
-    titleInput.addEventListener('keydown', (e) => {
-      if (e.key === 'Enter' && !e.shiftKey) e.preventDefault()
-    })
-    window.setTimeout(autosizeTitle, 0)
-    titleInput.focus()
-    if (this.isNew) titleInput.select()
-
-    const props = body.createDiv('pm-te-props')
-    renderTaskFormFields(props, {
-      task: this.task,
-      project: this.project,
-      plugin: this.plugin,
-      parentId: this.parentId,
-      setParentId: (id) => {
-        this.parentId = id
-      },
-      rerender: () => this.render(),
-      shownExtras: this.shownExtras
-    })
-
-    body.createEl('hr', { cls: 'pm-te-divider' })
-
-    const descSection = body.createDiv('pm-modal-section pm-modal-desc-section')
-    descSection.createEl('h4', { text: 'Description', cls: 'pm-modal-section-title' })
-
-    const descPreview = descSection.createDiv('pm-modal-desc-preview')
-    const descArea = descSection.createEl('textarea', { cls: 'pm-modal-description' })
-    descArea.placeholder = 'Add a description…'
-    descArea.value = this.task.description
-
-    const autoResize = () => {
-      const saved: [HTMLElement, number][] = []
-      let ancestor = descArea.parentElement
-      while (ancestor) {
-        if (ancestor.scrollTop > 0) saved.push([ancestor, ancestor.scrollTop])
-        ancestor = ancestor.parentElement
-      }
-      descArea.setCssProps({ '--desc-height': 'auto' })
-      descArea.setCssProps({ '--desc-height': descArea.scrollHeight + 'px' })
-      for (const [el, top] of saved) el.scrollTop = top
-    }
-
-    const hasContent = () => this.task.description.trim().length > 0
-    const sourcePath = this.task.filePath || this.project.filePath || ''
-
-    let descComp = new Component()
-    descComp.load()
-
-    const toggleCheckbox = (index: number) => {
-      let count = 0
-      this.task.description = this.task.description.replace(
-        /^([ \t]*[-*+] \[)([ x])(\])/gm,
-        (match, pre, state, post) => {
-          if (count++ === index) return pre + (state === ' ' ? 'x' : ' ') + post
-          return match
-        }
-      )
-      descArea.value = this.task.description
-      void renderPreview()
-    }
-
-    const attachCheckboxListeners = () => {
-      descPreview.querySelectorAll('input[type="checkbox"]').forEach((el, i) => {
-        const cb = el as HTMLInputElement
-        cb.removeAttribute('disabled')
-        cb.addEventListener('click', (e) => {
-          e.preventDefault()
-          toggleCheckbox(i)
-        })
-      })
-    }
-
-    // Electron silently drops file:// under the target="_blank" MarkdownRenderer emits.
-    const attachFileLinkHandlers = () => {
-      descPreview.querySelectorAll<HTMLAnchorElement>('a.external-link').forEach((a) => {
-        if (!a.href.startsWith('file://')) return
-        a.addEventListener('click', (e) => {
-          e.preventDefault()
-          activeWindow.open(a.href)
-        })
-      })
-    }
-
-    const renderPreview = async () => {
-      descComp.unload()
-      descComp = new Component()
-      descComp.load()
-      descPreview.empty()
-      await MarkdownRenderer.render(this.app, this.task.description, descPreview, sourcePath, descComp)
-      attachCheckboxListeners()
-      attachFileLinkHandlers()
-    }
-
-    const showEdit = (caret?: number) => {
-      descPreview.classList.add('pm-hidden')
-      descArea.classList.remove('pm-hidden')
-      descArea.value = this.task.description
-      window.setTimeout(() => {
-        autoResize()
-        descArea.focus()
-        if (caret !== undefined) descArea.setSelectionRange(caret, caret)
-      }, 0)
-    }
-
-    const showPreview = () => {
-      if (!hasContent()) return
-      void renderPreview()
-      descArea.classList.add('pm-hidden')
-      descPreview.classList.remove('pm-hidden')
-    }
-
-    descArea.addEventListener('input', () => {
-      this.task.description = descArea.value
-      autoResize()
-    })
-    descArea.addEventListener('blur', () => showPreview())
-
-    descArea.addEventListener('paste', (e) => {
-      const items = e.clipboardData?.items
-      if (!items) return
-      const attachments: { blob: Blob; name: string }[] = []
-      for (const item of Array.from(items)) {
-        if (item.kind === 'file' && item.type.startsWith('image/')) {
-          const file = item.getAsFile()
-          if (file) {
-            const stamp = new Date().toISOString().replace(/[:.]/g, '-')
-            const sub = (item.type.split('/')[1] || 'png').split('+')[0]
-            const ext = sub === 'jpeg' ? 'jpg' : sub
-            attachments.push({ blob: file, name: `Pasted-${stamp}.${ext}` })
-          }
-        }
-      }
-      if (attachments.length === 0) return
-      e.preventDefault()
-      void this.insertAttachments(descArea, attachments, autoResize)
-    })
-
-    descSection.addEventListener('dragover', (e) => {
-      if (!e.dataTransfer) return
-      if (!Array.from(e.dataTransfer.types).includes('Files')) return
-      e.preventDefault()
-    })
-
-    descSection.addEventListener('drop', (e) => {
-      const files = e.dataTransfer?.files
-      if (!files || files.length === 0) return
-      e.preventDefault()
-      if (descArea.classList.contains('pm-hidden')) {
-        showEdit()
-        descArea.selectionStart = descArea.selectionEnd = descArea.value.length
-      }
-      const attachments = Array.from(files).map((f) => ({ blob: f, name: f.name }))
-      void this.insertAttachments(descArea, attachments, autoResize)
-    })
-
-    this.noteSuggest?.destroy()
-    this.noteSuggest = new NoteLinkSuggest(this.app, descArea, (newValue) => {
-      this.task.description = newValue
-      autoResize()
-    })
-    this.noteSuggest.attach(descSection)
-
-    // Walk rendered text and markdown source in step, skipping source characters that
-    // produced no output, so a caret in the preview lands past the syntax around it.
-    const sourceOffsetOf = (renderedIndex: number) => {
-      const rendered = descPreview.textContent || ''
-      const src = this.task.description
-      const plain = (c: string) => (/\s/.test(c) ? ' ' : c)
-      let cursor = 0
-      for (let i = 0; i < renderedIndex && i < rendered.length; i++) {
-        const ch = plain(rendered[i])
-        while (cursor < src.length && plain(src[cursor]) !== ch) cursor++
-        cursor++
-      }
-      return Math.min(cursor, src.length)
-    }
-
-    const clickedSourceOffset = (e: MouseEvent) => {
-      const doc = descPreview.ownerDocument
-      const caret = doc.caretPositionFromPoint?.(e.clientX, e.clientY)
-      const node = caret?.offsetNode
-      if (!node || node.nodeType !== Node.TEXT_NODE || !descPreview.contains(node)) return undefined
-      const walker = doc.createTreeWalker(descPreview, NodeFilter.SHOW_TEXT)
-      let rendered = 0
-      let current = walker.nextNode()
-      while (current && current !== node) {
-        rendered += (current.textContent || '').length
-        current = walker.nextNode()
-      }
-      return current ? sourceOffsetOf(rendered + caret.offset) : undefined
-    }
-
-    descPreview.addEventListener('click', (e) => {
-      const target = e.target as HTMLElement
-      if (target.instanceOf(HTMLInputElement) && target.type === 'checkbox') return
-
-      const link = target.closest('a')
-
-      if (link) {
-        if (link.classList.contains('internal-link')) {
-          e.preventDefault()
-          e.stopPropagation()
-          const href = link.getAttribute('data-href') || link.getAttribute('href') || ''
-          this.saved = false
-          this.cancelled = false
-          this.close()
-          void this.app.workspace.openLinkText(href, sourcePath)
-          return
-        }
-        return
-      }
-
-      if (target.instanceOf(HTMLImageElement)) return
-
-      const selection = activeWindow.getSelection()
-      if (selection && !selection.isCollapsed && descPreview.contains(selection.anchorNode)) return
-
-      showEdit(clickedSourceOffset(e))
-    })
-
-    if (hasContent()) {
-      descArea.classList.add('pm-hidden')
-      void renderPreview()
-    } else {
-      descPreview.classList.add('pm-hidden')
-      window.setTimeout(autoResize, 0)
-    }
-
-    renderSubtasksPanel(body, this.task, this.plugin, this.plugin.store.configFor(this.project).statuses)
-
-    renderTimeTrackingPanel(body, this.task)
-
-    const footer = contentEl.createDiv('pm-te-footer')
-
-    if (!this.isNew && this.task.filePath) {
-      const filePath = this.task.filePath
-      const pathHint = footer.createSpan({ cls: 'pm-te-footer-path pm-te-copyable' })
-      const fileIcon = pathHint.createSpan({ cls: 'pm-te-footer-icon' })
-      setIcon(fileIcon, 'file-text')
-      pathHint.createSpan({ text: filePath })
-      setTooltip(pathHint, 'Copy file path')
-      pathHint.addEventListener(
-        'click',
-        safeAsync(async () => {
-          await navigator.clipboard.writeText(filePath)
-          new Notice('Copied file path')
-        })
-      )
-    }
-
-    footer.createDiv('pm-footer-spacer')
-
-    new ButtonComponent(footer).setButtonText('Cancel').onClick(() => {
-      this.cancelled = true
-      this.close()
-    })
-
-    const saveBtn = new ButtonComponent(footer)
-      .setButtonText(this.isNew ? 'Create (Shift+Enter)' : 'Save (Shift+Enter)')
-      .setCta()
-    let saving = false
-    const doSave = async () => {
-      if (saving) return
-      saving = true
-      try {
-        if (!this.task.title.trim()) {
-          titleInput.focus()
-          titleInput.classList.add('pm-input-error')
-          return
-        }
-        clearTitleError()
-        await this.persistTask()
-        this.saved = true
-        this.close()
-      } catch (err) {
-        if (err instanceof TaskFileNameConflictError) {
-          showTitleError(`A note named "${err.fileName}" already exists. Choose a different title.`)
-          return
-        }
-        console.error('[PM]', err)
-        new Notice('Something went wrong. Check the console for details.')
-      } finally {
-        saving = false
-      }
-    }
-
-    saveBtn.onClick(() => {
-      void doSave()
-    })
-
-    if (this.saveKeyHandler) this.modalEl.removeEventListener('keydown', this.saveKeyHandler)
-    this.saveKeyHandler = (e: KeyboardEvent) => {
-      if (e.key === 'Enter' && e.shiftKey) {
-        e.preventDefault()
-        void doSave()
-      }
-    }
-    this.modalEl.addEventListener('keydown', this.saveKeyHandler)
   }
 }
